+++ CHANGELOG.md
@@ -12,6 +12,7 @@ If you have used or are currently using this library and would like to upgrade, 
   * `ReactionMenu.remove_on_timeout()`
 * `RM|VM` Added parameter `reply` to the `start` method. Enables the menu message to reply to the message that triggered it
   * `ReactionMenu.start(..., reply: bool=False)`
+* `RM|VM` Added property `ReactionMenu.last_viewed`. Returns the last page that was seen by the user in the pagination process
 * `RM` Added factory methods for `ReactionButton`
   * `ReactionButton.back()` 
   * `ReactionButton.next()`